trackpoint-app/requirements.txt
numpy-stl
matplotlib
PyQt6
gmsh
pygmsh